Add music!

+++ src/Level.hpp
@@ -6,7 +6,7 @@ class Level {
 public:
     std::string m_levelName;
     std::string m_levelString;
-    int64_t m_audioTrack;
+    int m_audioTrack;
 
     static Level* fromGMD(std::string path);
     static Level* fromServers(int id);
